requirements.txt
# Web Framework ve Sunucu
fastapi==0.109.2
uvicorn[standard]==0.27.1

# Veritabanı
sqlalchemy==2.0.27
psycopg2-binary==2.9.9
aiosqlite==0.19.0

# Elasticsearch
elasticsearch==8.12.0

# GraphQL
strawberry-graphql==0.219.2

# Güvenlik ve Doğrulama
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Yardımcı Kütüphaneler
python-dotenv==1.0.1
pydantic==2.6.1
slowapi==0.1.9
Faker==22.6.0
numpy==1.26.1
jinja2==3.1.3
python-multipart==0.0.6
aiofiles==23.2.1

# AI ve NLP
openai==1.12.0
tiktoken==0.6.0